+++ lib/phoenix_ecto/html.ex
@@ -9,7 +9,7 @@ if Code.ensure_loaded?(Phoenix.HTML) do
         impl: __MODULE__,
         id: name,
         name: name,
-        errors: form_for_errors(changeset.errors),
+        errors: changeset.errors,
         model: model,
         params: params || %{},
         hidden: form_for_hidden(model),
@@ -51,7 +51,7 @@ if Code.ensure_loaded?(Phoenix.HTML) do
               impl: __MODULE__,
               id: id,
               name: name,
-              errors: form_for_errors(changeset.errors),
+              errors: changeset.errors,
               model: model,
               params: changeset.params || %{},
               hidden: form_for_hidden(model),
@@ -85,7 +85,7 @@ if Code.ensure_loaded?(Phoenix.HTML) do
               id: id <> "_" <> index_string,
               name: name <> "[" <> index_string <> "]",
               index: index,
-              errors: form_for_errors(changeset.errors),
+              errors: changeset.errors,
               model: model,
               params: changeset.params || %{},
               hidden: form_for_hidden(model),
@@ -236,17 +236,6 @@ if Code.ensure_loaded?(Phoenix.HTML) do
 
     defp form_for_method(%{__meta__: %{state: :loaded}}), do: "put"
     defp form_for_method(_), do: "post"
-
-    defp form_for_errors(errors) do
-      for {attr, message} <- errors do
-        {attr, form_for_error(message)}
-      end
-    end
-
-    defp form_for_error(msg) when is_binary(msg), do: msg
-    defp form_for_error({msg, count: count}) when is_binary(msg) do
-      String.replace(msg, "%{count}", to_string(count))
-    end
 
     defp warn_name(opts) do
       if name = Keyword.get(opts, :name) do